logo: Convert font to curve
Previously the `V` was still a font character, which isn't very clever, as
we'd need the right font to display the icon properly.

This patch converts the V to a curve to make it font independent.

Sadly the icon is slightly different from the previous version, but this
shouldn't be noticed by most eyes.

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>