design-proposals: Refactor and extend the Flavors API
KubeVirt's VirtualMachine API contains many advanced options for tuning
a virtual machine performance that goes beyond what typical users need
to be aware of. Users are unable to simply define the storage/network
they want assigned to their VM and then declare in broad terms what
quality of resources and kind of performance they need for their VM.

This design proposal sets of various changes to existing
VirtualMachineFlavor CRDs and the introduction of a new set of
VirtualMachinePreference CRDs to hopefully simplify the choices a user
has to make when defining a VirtualMachine to get the resources,
performance and guestOS characteristics they desire.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>