instancetype: Add ControllerRevision upgrade proposal
`VirtualMachines` referencing instance types or preferences have
`ControllerRevisions` created containing a point in time copy of the
referenced resource. See the user guide for more on this:

https://kubevirt.io/user-guide/virtual_machines/instancetypes/#versioning

While this allows KubeVirt to always produce the same
`VirtualMachineInstance` at run time regardless of changes to the
original resource it does currently mean that KubeVirt has to support
older versions of the `instancetype.kubevirt.io` CRDs for the lifetime
of these `VirtualMachines`.

This proposal outlines a solution to this by upgrading older
`ControllerRevisions` during a `VirtualMachine`  resync by the
`vm-controller`.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>