instancetype: Add ControllerRevision upgrade proposal
`VirtualMachines` referencing instance types or preferences have
`ControllerRevisions` created containing a point in time copy of the
referenced resource. See the user guide for more on this:

https://kubevirt.io/user-guide/virtual_machines/instancetypes/#versioning

While this allows KubeVirt to always produce the same
`VirtualMachineInstance` at run time regardless of changes to the
original resource it does currently mean that KubeVirt has to support
older versions of the `instancetype.kubevirt.io` CRDs for the lifetime
of these `VirtualMachines`.

This document proposes the introduction of two new controllers to
`virt-controller` to migrate the stashed `instancetype.kubevirt.io`
objects within these `ControllerRevisions` to their latest version,
hopefully allowing support for older deprecated versions to be dropped
from the codebase.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>