Revising proposal regarding configuration of machine types to be updated.
After some more discussion on the draft PR, we decided to modify the design
slightly. Rather than updating the machine type to a specific version, we will
be removing the machine type in VMs that need their machine type updated. Thus,
when the VM is then started/restarted, its machine type will be populated with
the default machine type defined in Kubevirt CR (which should always be a
supported machine type). Additionally, the user will need to specify the machine
type they wish to update by passing a regex for the machine type to the new
`--which-matches-glob` flag.

Link to discussion on draft PR: https://github.com/kubevirt/kubevirt/pull/10222#issuecomment-1689491133

Signed-off-by: prnaraya <prnaraya@redhat.com>