design-proposals: Allow custom guest CPU topologies within instancetypes
At present users of the `instancetype.kubevirt.io/v1beta1` API and CRDs
are able to define a CPU topology for a `VirtualMachine` through a
combination of vCPUs provided by the `guest` attribute of an instance
type and an optional topology provided by the `preferredCPUTopology`
attribute of a preference.

This design proposal aims to extend the API to allow for an optional
guest CPU topology to be provided by a referenced instance type instead
of a generic number of vCPUs.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>