network binding plugin - DownwardAPI for device info (#281)
Design proposal to extend the network binding plugin entry in the kubevirt
CR with an option for a downward API.
When specified, a downwardAPI from the virt-launcher pod to the binding
sidecar will be added.
Currently the only supported downward API is "device-info", exposing the
a map between the logical network name to its network status annotation
device info.

Signed-off-by: Alona Paz <alkaplan@redhat.com>