instancetypes: Extend preferSpread
The `preferSpread` `preferredCPUTopology` was introduced with KubeVirt
v1.1 and allows for vCPUs provided by an instance type to be spread
across guest visible sockets and cores using a configurable ratio that
defaults to 2.

This design proposal aims to extend this implementation by introducing
new configurables within a preference to control how vCPUs are spread
with the goal of allowing vCPUs to also be spread to threads.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>