update descheduler support design proposal
The descheduler support design proposal specifies that a
feature gate will be added.
The feature gate can be dropped, since no work is added other
than to add annotations.
The feature gate drop also has the advantage of saving work to
the cluster admins; with the feature gate, if a cluster admin
wants to enable the descheduler, they have to install it and
add the feature gate to KubeVirt.
Without the feature gate they only have to install it.

Also, this feature gate only enables a support for the
descheduler, which should always be enabled, and there are no
advantages to not have it.

Signed-off-by: fossedihelm <ffossemo@redhat.com>