design-proposal: VirtualMachineInstanceMigration - Live migration to specific nodes
Adding a design proposal to extend VirtualMachineInstanceMigration
object with an additional API to let a cluster admin
try to trigger a live migration of a VM injecting
on the fly an additional NodeSelectorTerm constraint.
The additional NodeSelectorTerm can only restrict the set
of Nodes that are valid target for the migration
(eventually down to a single host); in order to
achieve this,  all the `NodeSelectorRequirements`
defined on the additional `NodeSelectorTerm` set
on the VMIM object should be appended to
all the `NodeSelectorTerms` already defined on
the VM object.
All the affinity rules defined on the VM spec are still
going to be satisfied.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>