node-labeller rewrite
The node-labeller has deviated from its primary purpose of simply labeling nodes. It now also handles inferring host capabilities for the vmcontroller, a task unrelated to its core function. Most of the capabilities it gathers aren't relevant to node labeling, leading to an overcomplicated design and unnecessary coupling with the vmcontroller which hinders testing and extensibility of both components.

As such this design proposal seeks to decouple the two and simplify
node-labeller to focus on its original intended purpose while moving
its current inference logic to a separate virt-handler unit.

Signed-off-by: Adi Aloni <adialoni101@gmail.com>