design-proposals: Clean template and add feature lifecycle section
As discussed at the unconference for v1.5.0 [1] we should start
documenting the expected lifecycle of a given feature within our design
proposals before later moving this tracking to an issue.

This change cleans up the template, removing multiple top level
headings, adding whitespace around headings and adds a final section
defining the lifecycle of a feature.

[1] https://unconference.kubevirt.io/D62SngbSRXO1uhTAc8Ikqg#xpivarc-Feature-lifecycleFuture-of-proposals---1-1-1-1

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>