design-proposal: limit Live Migration initial permissions to cluster-admin
In kubernetes we have two distinct pre-defined
administrative roles:
- cluster-admin that is allowed to to perform any action on any resource
- admin that is allowed to read/write/amend almost all the resources in a namespace
We can assume that that granting to a developer/devops person
the admin role to a namespace that he "owns" is a common practice.

In order to enque o cancel a live migration request,
the user is supposed to create a VirtualMachineInstanceMigration object
that is a namespaced resource that lives in
the same namespace of the VM object.
Till now, a namespace admin is allowed to create/delete
VirtualMachineInstanceMigration objects in his namespaces.

But a live migration is also a critical feature for
cluster wide tasks such as upgrades, node drains and so on.
So today an unprivileged user can already easily delay a
cluster upgrade or a node drain simply continuously enqueuing
migrations or even block it by continuously deleting
VirtualMachineInstanceMigration objects as soon as they appear.

This proposal is about allowing, by default, only cluster-admin users
to create/delete VirtualMachineInstanceMigration objects
as a hardening measure (principle of least privilege).

A cluster-admin will still be allowed to grant the migration role
to selected users or groups.

Other actions like hotplugging that are only implicitly relying
on a live migration are not going to be affected by this change.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>