design-proposal: VirtualMachineInstanceMigration - Live migration to a named node (#320)
* design-proposal: VirtualMachineInstanceMigration - Live migration to specific nodes

Adding a design proposal to extend VirtualMachineInstanceMigration
object with an additional API to let a cluster admin
try to trigger a live migration of a VM injecting
on the fly an additional NodeSelectorTerm constraint.
The additional NodeSelectorTerm can only restrict the set
of Nodes that are valid target for the migration
(eventually down to a single host); in order to
achieve this,  all the `NodeSelectorRequirements`
defined on the additional `NodeSelectorTerm` set
on the VMIM object should be appended to
all the `NodeSelectorTerms` already defined on
the VM object.
All the affinity rules defined on the VM spec are still
going to be satisfied.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>

* Update migration-target.md

Added that this is for exceptions

Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>

* Make goals and user stories more explict

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>

* Address review comments

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>

---------

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>
Co-authored-by: Fabian Deutsch <fabiand@users.noreply.github.com>