sigs: Introduce sig-architecture
This commit introduces an Architecture SIG is also introduced to act as
a body to oversee the architectural consistency of the project,
including but not limited to API conventions, design principles and the
enhancement process. The latter being something that was informally
handled by sig-compute during v1.7.0 but should be moved elsewhere.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>